modelgen + single cuda_run implemented
