+++ dsl/TVB_testsuit/runthings
@@ -2,7 +2,7 @@
 
 rm error.er 
 rm output.out
-sed "s/bencharg/$1/" benchAll.sh | bsub
+sed -e "s/bencharg/${1}/;s/mdlrun/${2}/" benchAll.sh | sbatch
 #sleep 2
 
 while [ ! -f ./error.er ] ;
@@ -12,8 +12,8 @@ done
 
 if [ $1 == "regular" ]; 
 then
-  tail -f ./output.out | grep -i -E "INFO|WARNING|ERROR"
-  tail -f ./error.er
+  #tail -f ./output.out | grep -i -E "INFO|WARNING|ERROR"
+  tail -n 50 -f ./error.er
 else
-  tail -f ./error.er
+  tail -n 50 -f ./error.er
 fi
